update rb file name
