+++ lib/get_github_pub_keys.rb
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 require 'get_github_pub_keys/connection'
-require 'get_github_pub_keys/files'
+require 'get_github_pub_keys/file'
 require 'get_github_pub_keys/client'
 
 module GetGithubPubKeys
