+++ lib/get_github_pub_keys/client.rb
@@ -4,24 +4,20 @@ module GetGithubPubKeys
 
   class Client
     include Connection
-    include GetGithubPubKeys::Files
+    include GetGithubPubKeys::File
 
-    def initialize(user=nil) 
+    def initialize(user=nil)
       @user = user
     end
 
     def find_and_create( options = {} )
-      response = Connection.new(options).get("users/#{@user}/keys")   
+      response = Connection.new(options).get("users/#{@user}/keys")
       public_keys = response.body
 
-      # FIXME:response.raise_errorが正しく動作するのかを確認する
-      # 登録されていないUserだとfile_nameがとれない
       if public_keys.is_a? Array
-        public_keys.each do |public_key| 
-          GetGithubPubKeys::Files.create public_key["id"], public_key["key"]
-        end
+        public_keys.each { |public_key| GetGithubPubKeys::File.create public_key['id'], public_key['key'] }
       else
-        GetGithubPubKeys::Files.create public_keys["id"], public_keys["key"]
+        GetGithubPubKeys::File.create public_keys['id'], public_keys['key']
       end
     end
   end
